+++ pylinac/plan_generator/dicom.py
@@ -63,7 +63,6 @@ class Beam:
         self,
         plan_dataset: Dataset,
         beam_name: str,
-        beam_type: BeamType,
         energy: float,
         dose_rate: int,
         x1: float,
@@ -90,8 +89,6 @@ class Beam:
             The plan dataset. Used for dynamic links to other Sequences of the plan, such as Dose Reference and Tolerance Table.
         beam_name : str
             The name of the beam. Must be less than 16 characters.
-        beam_type : BeamType
-            The type of beam: dynamic or static.
         energy : float
             The energy of the beam.
         dose_rate : int
@@ -132,11 +129,6 @@ class Beam:
         if len(beam_name) > 16:
             raise ValueError("Beam name must be less than or equal to 16 characters")
 
-        if beam_type == BeamType.STATIC and len(mlc_positions) != 1:
-            raise ValueError(
-                f"Static beam can only have one MLC position change ({len(mlc_positions)})"
-            )
-
         if len(metersets) != len(mlc_positions):
             raise ValueError(
                 f"The number of meter sets ({len(metersets)}) "
@@ -151,8 +143,10 @@ class Beam:
                 "Cannot specify multiple gantry angles without a gantry direction"
             )
 
-        if isinstance(gantry_angles, Iterable) and beam_type == BeamType.STATIC:
-            raise ValueError("Cannot specify multiple gantry angles as a static beam")
+        gantry_is_static = len(np.unique(gantry_angles)) == 1
+        mlc_is_static = np.unique(np.array(mlc_positions), axis=0).shape[0] == 1
+        beam_is_static = gantry_is_static and mlc_is_static
+        beam_type = BeamType.STATIC if beam_is_static else BeamType.DYNAMIC
         self.plan_ds = plan_dataset
         self.ds = self._create_basic_beam_info(
             beam_name,
@@ -417,7 +411,6 @@ class HalcyonBeam(Beam):
         self,
         plan_dataset: Dataset,
         beam_name: str,
-        beam_type: BeamType,
         energy: float,
         dose_rate: int,
         machine_name: str,
@@ -438,8 +431,6 @@ class HalcyonBeam(Beam):
             The plan dataset. Used for dynamic links to other Sequences of the plan, such as Dose Reference and Tolerance Table.
         beam_name : str
             The name of the beam. Must be less than 16 characters.
-        beam_type : BeamType
-            The type of beam: dynamic or static.
         energy : float
             The energy of the beam.
         dose_rate : int
@@ -466,11 +457,6 @@ class HalcyonBeam(Beam):
         if len(beam_name) > 16:
             raise ValueError("Beam name must be less than or equal to 16 characters")
 
-        if beam_type == BeamType.STATIC and (
-            len(proximal_mlc_positions) != 1 or len(distal_mlc_positions) != 1
-        ):
-            raise ValueError("Static beam can only have one MLC position change.")
-
         if len(metersets) != len(proximal_mlc_positions) != len(distal_mlc_positions):
             raise ValueError(
                 f"The number of meter sets ({len(metersets)}) "
@@ -485,8 +471,17 @@ class HalcyonBeam(Beam):
                 "Cannot specify multiple gantry angles without a gantry direction"
             )
 
-        if isinstance(gantry_angles, Iterable) and beam_type == BeamType.STATIC:
-            raise ValueError("Cannot specify multiple gantry angles as a static beam")
+        gantry_is_static = len(np.unique(gantry_angles)) == 1
+        distal_mlc_is_static = (
+            np.unique(np.array(distal_mlc_positions), axis=0).shape[0] == 1
+        )
+        proximal_mlc_is_static = (
+            np.unique(np.array(proximal_mlc_positions), axis=0).shape[0] == 1
+        )
+        beam_is_static = (
+            gantry_is_static and distal_mlc_is_static and proximal_mlc_is_static
+        )
+        beam_type = BeamType.STATIC if beam_is_static else BeamType.DYNAMIC
         self.plan_ds = plan_dataset
         self.ds = self._create_basic_beam_info(
             beam_name,
@@ -1135,7 +1130,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         beam = Beam(
             plan_dataset=self.ds,
             beam_name=beam_name,
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=dose_rate,
             x1=x1,
@@ -1237,7 +1231,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         beam = Beam(
             plan_dataset=self.ds,
             beam_name=f"{beam_name} {bank}",
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=dose_rate,
             x1=x1,
@@ -1391,7 +1384,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         ref_beam = Beam(
             plan_dataset=self.ds,
             beam_name="DR Ref",
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=default_dose_rate,
             x1=roi_centers[0] - roi_size_mm / 2 - jaw_padding_mm,
@@ -1415,7 +1407,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         beam = Beam(
             plan_dataset=self.ds,
             beam_name=f"DR{min(dose_rates)}-{max(dose_rates)}",
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=default_dose_rate,
             x1=roi_centers[0] - roi_size_mm / 2 - jaw_padding_mm,
@@ -1584,7 +1575,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         ref_beam = Beam(
             plan_dataset=self.ds,
             beam_name=f"{beam_name} Ref",
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=default_dose_rate,
             x1=roi_centers[0] - roi_size_mm / 2 - jaw_padding_mm,
@@ -1608,7 +1598,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         beam = Beam(
             plan_dataset=self.ds,
             beam_name=beam_name,
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=default_dose_rate,
             x1=roi_centers[0] - roi_size_mm / 2 - jaw_padding_mm,
@@ -1706,7 +1695,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
             beam = Beam(
                 plan_dataset=self.ds,
                 beam_name=beam_name,
-                beam_type=BeamType.DYNAMIC,
                 energy=energy,
                 dose_rate=dose_rate,
                 x1=x1 - jaw_padding,
@@ -1860,7 +1848,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         beam = Beam(
             plan_dataset=self.ds,
             beam_name=beam_name,
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=max_dose_rate,
             x1=min(roi_centers) - roi_size_mm - jaw_padding_mm,
@@ -1884,7 +1871,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         ref_beam = Beam(
             plan_dataset=self.ds,
             beam_name=f"{beam_name} Ref",
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=max_dose_rate,
             x1=min(roi_centers) - roi_size_mm - jaw_padding_mm,
@@ -1988,7 +1974,6 @@ class TrueBeamPlanGenerator(PlanGenerator):
         beam = Beam(
             plan_dataset=self.ds,
             beam_name=beam_name,
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=dose_rate,
             x1=x1 - jaw_padding,
@@ -2173,7 +2158,6 @@ class HalcyonPlanGenerator(PlanGenerator):
         beam = HalcyonBeam(
             plan_dataset=self.ds,
             beam_name=beam_name,
-            beam_type=BeamType.DYNAMIC,
             energy=energy,
             dose_rate=dose_rate,
             gantry_angles=gantry_angle,