+++ docs/source/changelog.rst
@@ -10,6 +10,16 @@ Legend
 * :bdg-primary:`Refactor` denotes a code refactor; usually this means an efficiency boost or code cleanup.
 * :bdg-danger:`Change` denotes a change that may break existing code.
 
+v 3.37.0
+--------
+
+Plan Generator
+^^^^^^^^^^^^^^
+
+* :bdg-primary:`Refactor` The :class:`~pylinac.plan_generator.dicom.Beam` class no longer accepts the ``beam_type`` parameter.
+  A beam is considered dynamic if the gantry or mlc positions differ from the first control point for all control points.
+
+
 v 3.36.0
 --------